requirements.txt
psycopg2-binary
pyyaml
flask
pydantic 